Refactor dictionary implementation logic into a word list private API
It’s also an interesting refactoring here because now we only read the
dictionary once and random word uses the dictionary over and over again.
