+++ game/dictionary/lib/dictionary.ex
@@ -12,16 +12,8 @@ defmodule Dictionary do
       "contributors"
 
   """
+  alias Dictionary.WordList
 
-  def random_word do
-    word_list()
-    |> Enum.random()
-  end
-
-  def word_list do
-    "../assets/words.txt"
-    |> Path.expand(__DIR__)
-    |> File.read!()
-    |> String.split(~r/\n/)
-  end
+  defdelegate start(), to: WordList, as: :word_list
+  defdelegate random_word(word_list), to: WordList
 end
